put loading images in separate file from main
